main_agent/requirements.txt
a2a-sdk==0.3.10
agent-framework==1.0.0b251028
agent-framework-a2a==1.0.0b251028
agent-framework-azure-ai==1.0.0b251028
agent-framework-copilotstudio==1.0.0b251028
agent-framework-core==1.0.0b251028
agent-framework-devui==1.0.0b251028
agent-framework-lab==1.0.0b251024
agent-framework-mem0==1.0.0b251028
agent-framework-purview==1.0.0b251028
agent-framework-redis==1.0.0b251028
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
#aiosignal==1.4.0cd
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
azure-ai-agents==1.2.0b5
azure-ai-projects==1.1.0b4
azure-common==1.1.28
azure-core==1.36.0
azure-core-tracing-opentelemetry==1.0.0b12
azure-identity==1.25.1
azure-monitor-opentelemetry==1.8.1
azure-monitor-opentelemetry-exporter==1.0.0b44
azure-search-documents==11.6.0
azure-storage-blob==12.27.1
backoff==2.2.1
bcrypt==5.0.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
fastapi==0.121.0
fixedint==0.1.6
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
frozenlist==1.8.0
google-api-core==2.28.1
google-auth==2.42.1
googleapis-common-protos==1.71.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
jsonpath-ng==1.7.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mcp==1.20.0
mem0ai==1.0.0
microsoft-agents-activity==0.5.3
microsoft-agents-copilotstudio-client==0.5.3
microsoft-agents-hosting-core==0.5.3
ml_dtypes==0.5.3
msal==1.34.0
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.7.0
numpy==2.3.4
oauthlib==3.3.1
openai==1.109.1
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-asgi==0.59b0
opentelemetry-instrumentation-dbapi==0.59b0
opentelemetry-instrumentation-django==0.59b0
opentelemetry-instrumentation-fastapi==0.59b0
opentelemetry-instrumentation-flask==0.59b0
opentelemetry-instrumentation-psycopg2==0.59b0
opentelemetry-instrumentation-requests==0.59b0
opentelemetry-instrumentation-urllib==0.59b0
opentelemetry-instrumentation-urllib3==0.59b0
opentelemetry-instrumentation-wsgi==0.59b0
opentelemetry-proto==1.38.0
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-semantic-conventions-ai==0.4.13
opentelemetry-util-http==0.59b0
packaging==25.0
ply==3.11
portalocker==3.2.0
posthog==6.7.14
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.3
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
python-ulid==3.1.0
pytz==2025.2
#pywin32==311
PyYAML==6.0.3
qdrant-client==1.15.1
redis==6.4.0
redisvl==0.10.0
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.28.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sse-starlette==3.0.3
starlette==0.49.3
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0